Made changes after review
