+++ CHANGELOG.md
@@ -26,14 +26,12 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - When running `check_contracts` on a class with type aliases as type annotations for its attributes, the `NameError`
   that appears (which indicates that the type alias is undefined) is now resolved.
 - The default value of `pyta-number-of-messages` is now 0. This automatically displays all occurrences of the same error.
-- For the contract checking `new_setattr` function, any variables that depend only on `klass` are now defined in the
-  outer function, efficiency of code was improved, and the attribute value is now restored to the original value if the
-  `_check_invariants` call raises an error.
 - Added new function `validate_invariants` which takes in an object and checks that the representation invariants of the object are satisfied.
 - For the contract checking `new_setattr` function, the check for `ENABLE_CONTRACT_CHECKING` is now at the top of the
   function, any variables that depend only on `klass` are now defined in the outer function, efficiency of code was
   improved, and the attribute value is now restored to the original value if the `_check_invariants` call raises an
   error.
+- Added the file `conftest.py` to store `pytest` fixtures.
 
 ### Bug Fixes
