+++ tests/conftest.py
@@ -1,0 +1,11 @@
+import pytest
+
+import python_ta.contracts
+
+
+@pytest.fixture()
+def disable_contract_checking():
+    """Fixture for setting python_ta.contracts.ENABLE_CONTRACT_CHECKING = False."""
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = False
+    yield
+    python_ta.contracts.ENABLE_CONTRACT_CHECKING = True
